[REFACTOR]> move everything to AE namespace
